+++ performance_test/src/execution_tasks/publisher_task.cpp
@@ -50,7 +50,7 @@ void PublisherTask::run()
   perf_clock::time_point next_run = m_first_run + m_time_between_publish * m_loop_counter++;
   std::this_thread::sleep_until(next_run);
 
-  if (m_ec.is_zero_copy_transfer) {
+  if (m_ec.use_loaned_samples) {
     m_pub->publish_loaned(m_timestamp_provider, m_stats.next_sample_id());
   } else {
     m_pub->publish_copy(m_timestamp_provider, m_stats.next_sample_id());