+++ performance_test/CHANGELOG.md
@@ -7,6 +7,12 @@
 ### Changed
 - Renamed the `--dds-domain_id` CLI arg to `--dds-domain-id`
 - When `--dds-domain-id` is unspecified, fall back to the `ROS_DOMAIN_ID` environment variable
+- `--zero-copy` has been separated into two flags:
+  - `--shared-memory`: Enable shared-memory transfer in the plugin. This is meant to replace the
+    need to manually set runtime flags via `CYCLONEDDS_URI`, `APEX_MIDDLEWARE_SETTINGS`, etc.
+  - `--loaned-samples`: When publishing messages in the plugin, borrow loaned samples instead
+    of publishing by copy
+  - `--zero-copy` is now an alias for `--shared-memory --loaned-samples`
 
 ## 2.1.0 (2024/04/17)